pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "audiorepodigest"
version = "0.1.0"
description = "Automated email digests of the most important audio, music, DSP, and audio-AI GitHub repos on a configurable schedule."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
keywords = [
  "audio",
  "github",
  "digest",
  "dsp",
  "music",
  "speech",
  "audio-ai",
]
dependencies = [
  "croniter>=2.0.5",
  "httpx>=0.27.2",
  "jinja2>=3.1.4",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.6.1",
  "pyyaml>=6.0.2",
  "rich>=13.9.4",
  "tenacity>=9.0.0",
  "typer>=0.12.5",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11.2",
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.9",
]

[project.scripts]
audiodigest = "audiorepodigest.cli:app"

[project.urls]
Homepage = "https://github.com/cleider/AudioRepoDigest"
Repository = "https://github.com/cleider/AudioRepoDigest"

[dependency-groups]
dev = [
  "mypy>=1.11.2",
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.9",
]

[tool.hatch.build.targets.wheel]
packages = ["src/audiorepodigest"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "RUF"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
no_implicit_optional = true
namespace_packages = true
mypy_path = ["src"]

